pyproject.toml
[project]
name = 'GEOScore'
version = '0.1.0'
readme = 'README.md'
requires-python = '>=3.8'
license = {file = 'LICENSE'}
authors = [
    {name = 'Bokhonko Roman', email = 'nesanya1990@gmail.com'},
    {name = 'Mayorov Daniil', email = 'daniil.major123@gmail.com'},
    {name = 'Gnetov Ivan', email = 'gnetov.i@mail.ru'},
    {name = 'Tafincev Dmitry', email = 'tafincev05@gmail.com'},
    {name = 'Karabanov Denis', email = 'karabanovdenis2005@mail.ru'}
]
dependencies = ['fastapi[all]', 'uvicorn[standard]', 'requests', 'python-dotenv']

[build-system]
build-backend = 'mesonpy'
requires = ['meson-python']
